# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: ALVAS API
  version: 1.0.0
  description: API operativa para autenticacion, usuarios, propiedades, ventas, reportes e integraciones.
servers:
  - url: http://127.0.0.1:8787
    description: Wrangler local por defecto
  - url: https://alvas-api-staging.softwarelazana.workers.dev
    description: Staging
  - url: https://alvas-api.softwarelazana.workers.dev
    description: Production
security:
  - bearerAuth: []
...
paths:
...
  /usuarios:
    get:
      tags: [Usuarios]
      summary: Lista usuarios
      responses:
        "200":
          description: Lista de usuarios
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Success"
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/Usuario"
        "401":
          $ref: "#/components/responses/Error"
    post:
      tags: [Usuarios]
      summary: Crea usuario
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CrearUsuarioRequest"
      responses:
        "201":
          description: Usuario creado
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Success"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/Usuario"
        "400":
          $ref: "#/components/responses/Error"
...
  /propiedades:
    get:
      tags: [Propiedades]
      summary: Lista propiedades
      parameters:
        - name: leadId
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Lista de propiedades
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Success"
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/Propiedad"
    post:
      tags: [Propiedades]
      summary: Crea propiedad
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CrearPropiedadRequest"
      responses:
        "201":
          description: Propiedad creada
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Success"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/Propiedad"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    Error:
      description: Error de negocio, validacion, autenticacion o sistema
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    Success:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
          const: true
    ErrorResponse:
      type: object
      required: [success, message, code]
      properties:
        success:
          type: boolean
          const: false
        message:
          type: string
        code:
          type: string
        detalles:
          type: array
          items:
            type: object
            properties:
              path:
                type: string
              message:
                type: string
...
    CrearUsuarioRequest:
      type: object
      required: [username, nombre, clave, rol]
      properties:
        username:
          type: string
        nombre:
          type: string
        clave:
          type: string
        rol:
          type: string
          examples: [ADMIN, ASESOR]
...
    Usuario:
      type: object
      properties:
        id:
          type: string
        username:
          type: string
        nombre:
          type: string
        rol:
          type: string
        estado:
          type: string
        creadoEn:
          type: string
          format: date-time
        actualizadoEn:
          type: string
          format: date-time
    CrearPropiedadRequest:
      type: object
      required: [titulo, descripcion, precio]
      properties:
        titulo:
          type: string
        descripcion:
          type: string
        precio:
          type: number
          exclusiveMinimum: 0
        origen:
          type: string
        estado:
          type: string
        idLeadOrigen:
          type: string
        idClientePropietario:
          type: string
        captadaPorAsesorId:
          type: string
        asesorResponsableId:
          type: string
...
    Propiedad:
      type: object
      additionalProperties: true
      properties:
        id:
          type: string
        titulo:
          type: string
        descripcion:
          type: string
        precio:
          type: number
        estado:
          type: string
...